small but significant changes to the way that the code reads and stores host galaxy/halo properties.
in gizmo.io.Read.read_snapshots():
* changed input parameter assign_host_coordinates=True to assign_hosts=True, for brevity and to make it clear that the code can assign multiple hosts per simulation
* changed input parameter assign_host_principal_axes=False to assign_hosts_rotation=False, for similar reasons
* changed input parameter assign_host_orbits=False to assign_orbits=False, to make it clear that this computes the orbits wrt each host, not of each host itself

also, changed the way the particle catalog stores host properties, for added flexibility in future development and better namespace management. previously it stored as:
    part.host_positions, part.host_velocities, part.host_rotations
now it stores via a dictionary and it stores the axis ratios as well:
    part.host['position'], part.host['velocity'], part.host['rotation'], part.host['axis.ratios']

similarly, changed the naming convention for storing host properties across snapshots if reading the star_form_coordinates_600.hdf5 file. previously it stored as:
    part.host_positions_all, part.host_velocities_all, part.host_rotations_all
now it stores via a dictionary:
    part.hostz['position'], part.hostz['velocity'], part.hostz['rotation'], part.hostz['axis.ratios']