minor formatting and other clean ups from recent pull of Andrew Emerick's age tracer module.
changed way of supplying a subset of elemental abunances to read in read_snapshots(). previously, user supplied element_indices, but now, user supplies elements, a list of strings for which elemental abundances to read, which should be much more user friendly.

as part of this, revamped and simplified the elemental abunance pointer dictionary within the particle dictionary, but this should not affect users.